refactor(core): extract seed-fixture logic into reusable library
Split the seed-fixture script into:
- core/src/lib/seedFixtureProjection.ts: pure fixture->row projection,
  no DB dependency, safe to import from dry-run path
- core/src/lib/seedFixture.ts: upsert + edge creation + isFixtureSeeded
  gate, imported by both CLI and bootstrap

The CLI script now delegates. Idempotency, identity policy, and
dry-run-without-DATABASE_URL behavior are preserved.
